Fix latency definition
In Server mode latency is defined as the time from when the LoadGen was scheduled to pass a query to the SUT, to the time it receives a reply. For particularly bad implementations, the SUT can cause the LoadGen to get behind schedule and issue queries late. This rule change just legalizes what the LoadGen is already doing. The alternative, of starting the clock when the query is issued as opposed to when it was supposed to be issued, effectively rewards implementations for blocking the LoadGen's threads.