Update policy on caching
The caching rules are too strict. 

Consider a routine where there is some configuration state based on the shape of the input tensors to the network. MLPerf allows you to:

* instance the routine and inline every combination of shapes that will be seen in the validation set
* precompute the configuration state for every combination of shapes seen in the validation set

But it does not allow you to do that computation on the fly and maintain a cache of recently seen configurations, even though the latter is very likely the strategy used in production.