Remove section on availability
This section contains a definition of availability that is at odds with the definition in the main policies document, and is causing confusion, see https://github.com/mlcommons/inference_results_v0.7/issues/15. It also contains a statement about making experimental systems available for replication which is inconsistent with current practice, where you only need to make the system available if audited - this being covered by the audit rules.